+++ src/materials/Metallic.java
@@ -14,7 +14,7 @@ import utilities.Vector;
  * @author mld2443
  */
 public class Metallic extends Material {
-	private final float fuzz;
+	private final double fuzz;
 
 	/**
 	 * Construct a Metallic material. Can have a mirror finish, or a fuzziness
@@ -25,7 +25,7 @@ public class Metallic extends Material {
 	 * @param fuzz
 	 *            The fuzziness; 0.0 for mirror finish
 	 */
-	public Metallic(final Color color, final float fuzz) {
+	public Metallic(final Color color, final double fuzz) {
 		super(color, true);
 		this.fuzz = fuzz;
 	}
@@ -42,11 +42,11 @@ public class Metallic extends Material {
 	}
 
 	@Override
-	public Ray scatter(final Ray incoming, final Vector collision, final Vector normal, final float sceneIndex) {
+	public Ray scatter(final Ray incoming, final Vector collision, final Vector normal, final double sceneIndex) {
 		Vector reflected = Vector.reflect(incoming.direction, normal);
 
 		// Apply our fuzziness inside this if block
-		if (fuzz > 0.0f)
+		if (fuzz > 0.0)
 			reflected = Vector.sum(reflected, Vector.randomInUnitSphere().scale(fuzz));
 		
 		reflected = reflected.normalize();
@@ -55,10 +55,6 @@ public class Metallic extends Material {
 
 		// Metallic surfaces have the concept of a front and back side; If the
 		// ray hits the front, it reflects
-		//if (Vector.dot(bounce.direction, normal) > 0)
 		return bounce;
-
-		// If the ray hits the back side, it is absorbed.
-		//return null;
 	}
 }
